Create macro to dry up form section definitions
smartcitiesdata/smartcitiesdata#722
